Changed threshold to accept decimals so you get more precise control of the threshold

+++ Cronus Zen Message Screen Studio/DrawTextWindow.xaml.cs
@@ -41,7 +41,7 @@ namespace CronusZenMessageScreenStudio
             WhiteOnBlack = true;
         }
 
-        public int Threshold { get; set; }
+        public double Threshold { get; set; }
         public int MarginTop { get; set; }
         public int MarginBottom { get; set; }
         public int MarginLeft { get; set; }
