+++ Cronus Zen Message Screen Studio/LoadImageWindow.xaml.cs
@@ -33,7 +33,7 @@ namespace CronusZenMessageScreenStudio
             Position = ImageProcessor.Positions.Center;
         }
 
-        public int Threshold { get; set; }
+        public double Threshold { get; set; }
         public int MarginTop { get; set; }
         public int MarginBottom { get; set; }
         public int MarginLeft { get; set; }
